COT Wrapper
